balance : increase minimum pay for everyone. (#3519)
# About the pull request

rifleman shouldn't be so poor they
can't even afford to buy a pack of the
cheapest cigarette and a zippo...
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
everyone in the marines should have enought money to buy atleas a pack
of the cheapest cigarette or cigar and a zippo.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: increase minimum paygrade for every marines
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>