HPR is cast into the void (#3534)
# About the pull request

This PR removes the HPR from all vendors other than requisitions.

# Explain why it's good for the game

Ammo economy and scarcity of special weapons is integral to the CM
marine experience.

The availability and increasing universality of the HPR goes completely
counter to that.

Honest to god, this is a marine buff.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: HPR is now only found in requisitions.
/:cl: