Nerfs Lobby Music (#3541)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Reduces default volume of Lobby Music and makes it stop on joining the
game.

Unfortunately have to reduce volume because the volume preferences are
not saved.

Shamelessly untested

# Explain why it's good for the game
Less intrusive, and as with my 2 previous PRs on the topic, should
discourage new people from just closing the window right away.


# Changelog
:cl:
fix: Lobby Music is a bit quieter and should now stop properly when you
spawn in.
/:cl: