Admin Global Blurb (#3550)
# About the pull request
Gives admins and maintainers a button to send a global blurb message.
This sends to all players with a connected client.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Sometimes we really need to get info to everyone and they don't always
pay attention to the chat window where we're putting announcements. Also
useful for system messages like planned server outages.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Adds an admin button for global blurb alerts. (Visual on game
screen)
/:cl: