Removes fourth awkward LZ2 Soro communication relay (#3566)
# About the pull request

This PR removes the LZ2 Soro communication relay that is right next to
an LZ1 communication relay.

# Explain why it's good for the game

Given some of the coming objective changes it is too strong to have them
right next to each other.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
del: Removes fourth awkward LZ2 Soro communication relay
/:cl: