Refactor many object to remove code from DMM to DM code. (#3396)
# About the pull request
create many subtype to clean dmm code
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
because making sub-type is good for now and the future...

# Testing Photographs and Procedure
pass the tests server launched fine..
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: create many sub-type for item to clean the code.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>