fixes ridged armor not appearing properly: or, how i learned to regret my own actions (#3580)
# About the pull request
Oops! Even more armor prefs bugs
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

![image](https://github.com/cmss13-devs/cmss13/assets/70115628/dfe5a961-0963-4c84-a98c-be5392de801f)
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: fixed a bug with armor prefs
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>