Escape pod crash land chance (#3540)
# About the pull request

This PR makes evacuating in an escape pod have a 50% chance to crash
land to the surface.

This PR fixes already launched escape pods launching again when
evacuation is complete aboard the Almayer.

# Explain why it's good for the game

This is a part of a grander scale idea to create objectives during
hijack that will effect the chances of successful escape. The
probability of crashing will be tied directly into these systems when
complete. In the mean time, I think it is interesting to have a chance
to crash land.

Bug bad.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Escape pods now have a 50% chance to crash land to the surface
fix: Escape pods will no longer relaunch when evacuation is complete
/:cl: