Huggers no longer lose health over time on weeds (#3603)
From what I can tell the damage is meant to effect huggers outside of
weeds so it is more of a QoL for huggers to make the role more bearable
to play on the weeds.

If it is meant to effect huggers on weeds too then I'd like to challenge
that idea because I find that part more of a chore vs a fun gameplay
aspect.

:cl: Hopek
balance: Huggers no longer lose health over time on weeds.
/:cl:

---------

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>