Fix Restore some object back to original to keep them compatible with older map (#3607)
# About the pull request
Restore some object back to original to keep them compatible with older
map that people are working on off the main maps.
it was a miss on my part and we have no reason to penalize people that
are working on maps off the grid...

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Because i don't want to piss off mapper for no good reason or give a
reason for people to give up project.
i don't want to create issues i want to solve them.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: restore some object back to original to keep them compatible with
older map.
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>