Smartgun battery (#3588)
# About the pull request

This PR removes the smartgun powerpack. You no longer need it for the
smartgun.

This PR adds a smartgun specific battery that you use with the smartgun
itself.


![Drakehtm56](https://github.com/cmss13-devs/cmss13/assets/8919187/f0ca314d-8a68-4bfd-8cdc-49930e9c77bc)


![Smart_Gun_arm](https://github.com/cmss13-devs/cmss13/assets/8919187/47ac63bf-8b1c-45b5-9ea3-e639f2afe88d)

The smartgun harness now blocks back slot.

# Explain why it's good for the game

Getting closer to the lore is good. Forcing more scarcity/attrition via
power on smartgunners is good for game health.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow, Triiodine
add: Added a smartgun specific battery that you use with the smartgun
itself.
del: Removed the smartgun powerpack. You no longer need it for the
smartgun.
/:cl: