Weed Addition To Trijent, Fix Larva Issue (#3638)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Adds a weed node to a secure cell adjacent to a xeno larva spawn
location on Trijent. If a player is unlucky there is a chance they'll
spawn inside the containment cell. The weed will allow this player to
grow from Larva to a T1 and break out on their own.

Fixes #3557 

# Explain why it's good for the game

Don't want people being trapped in a cell.


# Testing Photographs and Procedure


# Changelog

:cl:
fix: Larva who spawn in containment in Trijent will now have weeds to
grow on.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>