OT/Req delivery system repair (#3489)
# About the pull request
This is a simple fix for the delivery system between the OT workshop and
Requisitions. Previously the system has been either forgotten, left on a
to-do list somewhere, or not noticed entirely. I have added a
teleporting pipe connection from OT workshop to Req using the same
system as the one in Research, as well as a special sorting T-junction
to ensure items move correctly in the pipes.

I have tested the change and all appears to be working as intended
# Explain why it's good for the game
The OT/Req delivery system was a great concept and certainly an
excellent quality of life upgrade for the two distant departments.
However, since its initial addition, it has sat as an incomplete system
within the map files. With this quick and harmless fix, we can have this
great feature properly on its feet
# Photo Overview
<details>


![image](https://github.com/cmss13-devs/cmss13/assets/31124786/c17e1468-6459-4cb4-b749-217294328bb6)
_The OT side of the teleporting pipe system (no, this was not here
before)_


![image](https://github.com/cmss13-devs/cmss13/assets/31124786/a77a49f9-2961-432c-a589-1835dfe99e51)
_The new Req side of the pipe system with new custom T-junction sorter
(door was removed for ease of viewing)_

</details>

# Changelog
:cl:
fix: Completed the teleporting pipeline delivery system between OT
workshop and Requisitions
/:cl: