Fix XM88 streak end notification (#3682)
# About the pull request
XM88 has a notification on the hit streak end, but the end user
currently never sees it except for when the gun is dropped.

That's because it used visible_message and those don't work with things
in hand. So, you would only see it when you drop the gun. Also I've
changed it so that when the gun is dropped that message only appears
when there was an actual streak going.

# Explain why it's good for the game

REEEEE?!

# Changelog
:cl: sg2002
fix: XM88 now correctly notifies the user on the hit streak end.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>