Xeno Candidate Queue (#3636)
# About the pull request

This PR changes it so get_alien_candidates is normally a sorted list
based generally on timeofdeath in ascending order and larva spawns now
pull candidates out of that list in order. That means that an observer
that has yet to play, or was the first to die, and still meets all other
criteria to become a larva (not AFK, has preferences set to become
larva, has been dead long enough, etc.) will be chosen before others.

Playing as a facehugger (dying or hugging successfully) or dying in an
admin z-levels (such as thunderdome) also do not affect the value used
for how you are sorted in the queue. So you should be able to freely
play in either of those situations without losing your relative spot in
the queue. Of course its not going to nab you out of whatever mob you
are playing, but when you are a ghost again the same values will be used
to sort you in the queue.

Since people may enter and leave the queue each time the queue is
checked, your place in the queue may go up or down. It is just a
snapshot of that moment where you are. If you missed your queue message,
or haven't gotten one yet, the join xeno action will now display the
last message for you.

# Explain why it's good for the game

Picking candidates randomly is okay, but it would be more fair to give
privilege to those that have been waiting longer - especially those who
have yet to play.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Example of queen death code (since that was the most altered):

![larva](https://github.com/cmss13-devs/cmss13/assets/76988376/48211b1e-7556-4a41-879f-1dfbfabe8704)

</details>

# Changelog
:cl: Drathek
add: The selection to become a xeno larva is now based on timeofdeath
rather than random and also sends a message to all candidates when the
queue moves. Playing as a facehugger or on admin z-levels (thunderdome)
will not affect your relative place in queue. The join xeno action will
also display the last queue message for you when you allow xeno
candidacy.
fix: Fixed gibbed humans not properly setting their timeofdeath when on
an admin z-level.
fix: Fixed a bad del on shuttle doors when a shuttle deletes (such as
intoTheSunset).
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>