Fix erronous flamer_tank is empty messages and add sfx to empty canister verb (#3623)
# About the pull request

This PR is a followup to #3357 to fix an incorrect message discovered
when reviewing that PR as well as to make the existing empty_reagents
verb behave the same as the empty_reagents verb added in that PR.

# Explain why it's good for the game
The messages sent to the user should be accurate, and verbs should have
consistent behavior.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


[Refill](https://github.com/cmss13-devs/cmss13/assets/76988376/50b1cbca-ac0f-4856-9e27-8d1004c722f0)

</details>


# Changelog
:cl: Drathek
fix: Fixed erroneous is empty messages when inserting tanks into a
flamer pack.
soundadd: Added the refill sfx when using the empty canister verb on
flamer tanks.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>