Fix larva doubling when banished and returned to core (#3708)
# About the pull request

This PR fixes an oversight brought up by TheGamer01 where banishing
larva can double stored larva because banishment rewards it on death,
and if there is still a body that body can also be returned to core. Now
returning banished larva just has the same to_chat and effect as regular
larva, but there is no additional larva granted.

Also note that readmitting a dead xeno doesn't change the banished
status on the mob (if you want to test be sure to set the ckey of the
mob to something other than you).

# Explain why it's good for the game

No exploiting banishment please.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![larba](https://github.com/cmss13-devs/cmss13/assets/76988376/1e8c40bf-66bd-4a65-9e8c-a7793aa886e6)

</details>


# Changelog
:cl: Drathek
fix: Fix a possibility of banished larva refunding double the larva
/:cl: