Makes Predator ages fit with our lore (#3688)
# About the pull request
Makes the min and max pred ages fit more suitably with our
https://cm-ss13.com/wiki/Hunters#Biology
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Consistency is good.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Changed min and max Predator ages from 20 - 10000 to 175 - 3000
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>