Allow Mentors to Eavesdrop on Mhelps (#3720)
# About the pull request

This PR grants mentors the ability to see other responses to mentor help
messages, and fixes the double messaging if you are responder to the
mhelp.

# Explain why it's good for the game

Mentors can learn from other mentors too!

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl: Drathek
admin: Mentors can now eavesdrop on mentor messages, and the responder
to a mhelp doesn't get double logging from eavesdropping.
/:cl: