Liaison drip (#3620)
# About the pull request

This PR adds three new outfits to the CL's office.

This PR adds a new snow jacket for the CL.

See the icon checks in the "checks" tab to see the sprites.

# Explain why it's good for the game

Drip good. (sometimes)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow, Thwomper
add: Added three new uniforms and a snow jacket for the CL.
/:cl: