Fix Deterred Crashsite Offset (#3717)
# About the pull request

This PR fixes an oversight where a deterred crashsite does not get
positioned the same as a non-deterred one.

# Explain why it's good for the game

Fixes #3593 or atleast should much better mitigate it making it so a
deterred crashsite positions the same as a non-deterred.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl: Drathek
fix: Fixed the crashsite offset for a hijack shuttle that gets deterred
by the MGAD System
/:cl: