Big Red Liaison Survivor (#3746)
# About the pull request
Adds a Corporate Liaison survivor to Solaris Ridge. Tested it, was able
to spawn in and select it via the role category selection thing.
# Explain why it's good for the game
more options for roleplay are good, the map didn't have a survivor for
the corporate category, plus it makes sense to have one because of the
map's lore + the new liaison drip gets more use
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: IowaPotatoFarmer
add: Solaris Ridge now has a Corporate Liaison survivor.
/:cl: