Reduces the Burst timer from 10 minutes to 7.5 minutes (#3748)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
Reduces the Burst timer from 10 minutes to 7.5 minutes

# Explain why it's good for the game
Being Hugged is something that should be a life threat situation .
currently you can see most of marines getting hugged and not even going
back into the almayer because of the extremely long time they have to
get back , combined with stasis bag being hugged is as most a minor
anoyance for marines instead of a game ending threat.

PD Future implementation : a medicine made by chemistry that makes larva
go on stasis winout needing the bag. ( like super peri)

NOTE : this also affects HIVE burst timer . reducing it to 5 minutes
from 6.6 minutes but it can be reverted if it proves to be too quick.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: reduced Larva Burst time from 10 minutes to 7.5 minutes
refactor: changed 1 letter vars.
/:cl: