Lurker no longer decloaks if they pounce onto a dead guy. (#3724)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Doesn't make sense to lose your cloak for pouncing at a dead guy.

They'll still decloak if it's a living guy hiding under a dead guy.

(This only happens if your pounce ends on a tile where there's a dead
guy)

# Explain why it's good for the game
Inconsistent at best. Bad

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Lurkers no longer lose their pounce if they happen to end their
pounce on a tile with a dead human
/:cl: