People with lisp now lisp on radio (#3760)
# About the pull request

The lisp trait did not affect a flag for speech, making it so lisping
does not occur over radio, this should make it so what they say is
transmitted properly.

# Explain why it's good for the game

Fixes an oversight and allows for lisp speech to be handled by the
handle_speech_problems method. Also realism or something

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/91219575/7526b49e-ff74-4979-8024-71768501e43a)
Poor marine trying to alert command
</details>


# Changelog
:cl:
fix: lisps now occur over radio as well
refactor: Relevant variables relating to speech problems now use true
and false instead of 1 and 0
/:cl: