Dropship Missile Refactor (#3732)
# About the pull request
Refactors Dropship Missiles to not use Spawn().
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
While proposing a new CAS feature, this part of the code uses a function
we shouldn't be using. With a suggestion from Harry I simply swapped out
the parts that we could replace without having to do a bigger overhaul.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: refactored dropship_ammo so that missiles don't use Spawn()
/:cl: