Fix time of death persistence for larva queue (#3781)
# About the pull request

This PR moves the value used to sort players in the larva queue from the
client to player_details. In testing reconnecting a client seemed to
work because I didn't give enough time for the client to delete, but now
it should always persistent for the round.

# Explain why it's good for the game

Larva queue is supposed to actually give players who have yet to play
the priority, not players who have died, and disconnect long enough.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl: Drathek
fix: Fixed a persistence problem for the time of death value used for
the larva queue.
fix: Fixed brainmobs (human gibbing) not getting a time of death value.
/:cl: