Fountain pen (#3727)
Adds a Luxurious fountain pen to the game. Currently spawns on the right
ear of high ranking command staff as well as the corporate liaison.

What makes this pen special besides looking fancy?

![dreamseeker_ayCdkAmUBg](https://github.com/cmss13-devs/cmss13/assets/24533979/a5e3638f-2544-4bbd-a9da-6a4ffeca9f02)

![dreamseeker_bdX1Mp6lOy](https://github.com/cmss13-devs/cmss13/assets/24533979/28bbae7d-d364-42af-8a9c-7dce80f56840)

New colors to write with included with this amazon pen!! I think this
will enable deeper RP of having something be signed with an X color
meaning it was signed by command staff (or someone posing as command).


To-do that I haven't done yet because I haven't found the code and if I
don't do this before it is merged I'll just make a new PR for it so
don't let it block a merge:

- [x] Add a mob sprite to add these on to people's helmets so that a
marine can look fancy and decorated by receiving a pen. Sprites exist
for this just haven't implemented it yet.

:cl: Hopek, Gul Dreggrod
add: Adds a Luxurious fountain pen to the game. Currently spawns on the
right ear of high ranking command staff. Sprited by Gul Dreggrod.
/:cl: