Xeno structures/weeds become forsaken after hijack (#3685)
# About the pull request

This PR makes groundside xeno structures and weeds become forsaken after
hijack.

With blessing, finishing this up for the author of this:
https://github.com/cmss13-devs/cmss13/pull/3630

# Explain why it's good for the game

It is annoying to try and fight through all traps and such to go and do
stuff as a forsaken xeno.

I would just hardcode ally but some people enjoy the XvX that sometimes
occurs.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Newyearnewme, Morrow
add: Xeno structures/weeds now become forsaken after hijack
/:cl:

---------

Co-authored-by: Asbestos Gremlin <etgcollege@gmail.com>
Co-authored-by: NewyearnewmeUwu <70115628+NewyearnewmeUwu@users.noreply.github.com>
Co-authored-by: fira <loyauflorian@gmail.com>