Removes fountain pens from gear sets (#3800)
# About the pull request

This PR removes fountain pens from gear sets.

# Explain why it's good for the game

This is a flavor item and a preference item. It should not be in
non-specialized gear presets much less in ear slots.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
del: Removes fountain pens from gear sets
/:cl: